Add to cart page created

+++ Frontend/services/carServices.js
@@ -2,10 +2,10 @@ import axios from "axios";
 
 const API = axios.create({ baseURL: "http://localhost:10000" });
 
-export const getAllCars = () => API.get("/api/cars");
-export const getCarById = (id) => API.get(`/api/cars/${id}`);
-export const createCar = (data) => API.post("/api/cars", data);
-export const updateCar = (id, data) => API.put(`/api/cars/${id}`, data);
-export const deleteCar = (id) => API.delete(`/api/cars/${id}`);
+export const getAllCars = () => API.get("/api/admin/cars");
+export const getCarById = (id) => API.get(`/api/admin/cars/${id}`);
+export const createCar = (data) => API.post("/api/admin/cars", data);
+export const updateCar = (id, data) => API.put(`/api/admin/cars/${id}`, data);
+export const deleteCar = (id) => API.delete(`/api/admin/cars/${id}`);
